Refactor to lib+binary and add workspace
Workspace is created for the purpose of adding
rav1e-worker, which will act as a server for
distributed encoding.
rav1e-by-gop remains the client and retains
all existing functionality.

Splits rav1e-by-gop into a lib and bin section,
allowing binary-only components to be excluded
from rav1e-worker's dependency and keep the
rav1e-worker build times low.

Adds an empty project for rav1e-worker
(only Cargo.toml is filled out).
