+++ Cargo.toml
@@ -1,36 +1,5 @@
-[package]
-name = "rav1e-by-gop"
-version = "0.2.0"
-authors = ["Josh Holmer <jholmer.in@gmail.com>"]
-edition = "2018"
-license = "MIT"
-description = "A tool to parallel encode GOP segments of a video with rav1e"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0"
-av-scenechange = { version = "0.4", default-features = false, features = ["serde"] }
-clap = "2"
-console = "0.11.2"
-crossbeam-channel = "0.4"
-crossbeam-utils = "0.7"
-dialoguer = "0.5"
-env_logger = "0.7.1"
-indicatif = "0.14"
-itertools = "0.9"
-ivf = "0.1"
-log = "0.4"
-num_cpus = "1"
-rav1e = { git = "https://github.com/xiph/rav1e", default-features = false, features = ["asm"] }
-rmp-serde = "0.14"
-serde = { version = "1", features = ["derive"] }
-systemstat = "0.1"
-thiserror = "1"
-threadpool = "1"
-v_frame = "0.1"
-# Forked to work around lifetimes issues with sending the decoder across threads
-y4m = { version = "0.5", path = "y4m/" }
+[workspace]
+members = ["rav1e-by-gop", "rav1e-worker"]
 
 [profile.dev]
 opt-level = 1
@@ -38,7 +7,6 @@ opt-level = 1
 [profile.release]
 codegen-units = 1
 lto = true
-debug = true
 
 # Temporary workaround until rav1e 0.3.2 is released with v_frame support
 [patch.crates-io]
