+++ rav1e-by-gop/src/bin/decode.rs
@@ -9,6 +9,7 @@
 // PATENTS file, you can obtain it at www.aomedia.org/license/patent.
 
 use rav1e::prelude::*;
+use rav1e_by_gop::VideoDetails;
 use std::io::{self, Read};
 use thiserror::Error;
 use y4m::Decoder;
@@ -27,29 +28,6 @@ pub enum DecodeError {
     IoError(io::Error),
     #[error("Memory limit exceeded")]
     MemoryLimitExceeded,
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct VideoDetails {
-    pub width: usize,
-    pub height: usize,
-    pub bit_depth: usize,
-    pub chroma_sampling: ChromaSampling,
-    pub chroma_sample_position: ChromaSamplePosition,
-    pub time_base: Rational,
-}
-
-impl Default for VideoDetails {
-    fn default() -> Self {
-        VideoDetails {
-            width: 640,
-            height: 480,
-            bit_depth: 8,
-            chroma_sampling: ChromaSampling::Cs420,
-            chroma_sample_position: ChromaSamplePosition::Unknown,
-            time_base: Rational { num: 30, den: 1 },
-        }
-    }
 }
 
 pub(crate) fn get_video_details<R: Read>(dec: &Decoder<R>) -> VideoDetails {
